kube_core

Module cel

CEL validation for CRDs

Structs

Rule

Rule is a CEL validation rule for the CRD field

Enums

ListMerge

ListType represents x-kubernetes merge strategy for list.

MapMerge

MapMerge represents x-kubernetes merge strategy for map.

MergeStrategy

MergeStrategy represents set of options for a server-side merge strategy applied to a field.

Message

Message represents CEL validation message for the provided type

Reason

Reason is a machine-readable value providing more detail about why a field failed the validation.

StructMerge

StructMerge represents x-kubernetes merge strategy for struct.

Functions

merge_properties

Merge schema properties in order to pass overrides or extension properties from the other schema.

merge_strategy

Merge strategy takes schema and applies a set of merge strategy x-kubernetes rules to it, such as “x-kubernetes-list-type” and “x-kubernetes-list-map-keys”.

merge_strategy_property

Merge strategy property mutates property under property_index of the schema with the provided set of merge strategy rules.

validate

Validate takes schema and applies a set of validation rules to it. The rules are stored on the top level under the “x-kubernetes-validations”.

validate_property

Validate property mutates property under property_index of the schema with the provided set of validation rules.