Crate kube_core

Crate with types and traits necessary for interacting with the Kubernetes API

This crate provides the minimal apimachinery necessary to make requests to the kubernetes API.

It does not export export a client, but it also has almost no dependencies.

Everything in this crate is re-exported from kube (even with zero features) under kube::core.

Re-exports

pub use duration::Duration;

pub use dynamic::ApiResource;

pub use dynamic::DynamicObject;

pub use crd::CustomResourceExt;

pub use gvk::GroupVersion;

pub use gvk::GroupVersionKind;

pub use gvk::GroupVersionResource;

pub use metadata::PartialObjectMeta;

pub use metadata::PartialObjectMetaExt;

pub use metadata::TypeMeta;

pub use object::NotUsed;

pub use object::Object;

pub use object::ObjectList;

pub use request::Request;

pub use response::Status;

pub use watch::WatchEvent;

Modules

conversion

Contains types useful for implementing custom resource conversion webhooks.

crd

Traits and tyes for CustomResources

discovery

Type information structs for API discovery

duration

Kubernetes Durations.

dynamic

Contains types for using resource kinds not known at compile-time.

gvk

Type information structs for dynamic resources.

metadata

Metadata structs used in traits, lists, and dynamic objects.

object

Generic object and objectlist wrappers.

params

A port of request parameter *Optionals from apimachinery/types.go

request

Request builder type for arbitrary api types

response

Generic api response types

subresource

Request builder types and parameters for subresources

util

Utils and helpers

watch

Types for the watch api

Structs

ClusterResourceScope

Indicates that a Resource is cluster-scoped.

DynamicResourceScope

Indicates that a Resource is of an indeterminate dynamic scope.

ErrorResponse

An error response from the API.

ListMeta

ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

NamespaceResourceScope

Indicates that a Resource is namespace-scoped.

ObjectMeta

ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

SubResourceScope

Indicates that a Resource is neither cluster-scoped nor namespace-scoped.

Enums

Version

Version parser for Kubernetes version patterns

Traits

Resource

An accessor trait for a kubernetes Resource.

ResourceExt

Helper methods for resources.

ResourceScope

The scope of a Resource.