Crate kube_core

Crate with types and traits necessary for interacting with the Kubernetes API

This crate provides the minimal apimachinery necessary to make requests to the kubernetes API.

It does not export export a client, but it also has almost no dependencies.

Everything in this crate is re-exported from kube (even with zero features) under kube::core.

Re-exports

pub use dynamic::ApiResource;

pub use dynamic::DynamicObject;

pub use crd::CustomResourceExt;

pub use gvk::GroupVersion;

pub use gvk::GroupVersionKind;

pub use gvk::GroupVersionResource;

pub use metadata::TypeMeta;

pub use object::NotUsed;

pub use object::Object;

pub use object::ObjectList;

pub use request::Request;

pub use response::Status;

pub use watch::WatchEvent;

Modules

crd

Traits and tyes for CustomResources

discovery

Type information structs for API discovery

dynamic

Contains types for using resource kinds not known at compile-time.

gvk

Type information structs for dynamic resources.

metadata

Metadata structs used in traits, lists, and dynamic objects.

object

Generic object and objectlist wrappers.

params

A port of request parameter *Optionals from apimachinery/types.go

request

Request builder type for arbitrary api types

response

Generic api response types

subresource

Request builder types and parameters for subresources

util

Utils and helpers

watch

Types for the watch api

Structs

ErrorResponse

An error response from the API.

ListMeta

ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.

ObjectMeta

ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.

Enums

Error

Core error types.

Traits

Resource

An accessor trait for a kubernetes Resource.

ResourceExt

Helper methods for resources.

Type Definitions

Result

Convient alias for Result<T, Error>